requirements_dominic.txt
anthropic>=0.40.0
apscheduler>=3.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
tweepy>=4.14.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0
pytz>=2024.1
lxml>=4.9.0
python-dotenv>=1.0.0
